the delta and min_ver flags are now working properly, print statements yet to be removed

+++ kmc_core/splitter.h
@@ -47,7 +47,7 @@ class CSplitter
 
 	uint32 kmer_len;
 	uint32 minimizerVersion;
-	uint32 delta_val;
+	double delta_val;
 	uint32 window_len; // Souvadra's addition
 	//uint32 prefix_len;
 	uint32 signature_len;
@@ -166,7 +166,7 @@ template <typename COUNTER_TYPE> class CWSmallKSplitter {
 	uint64 total_kmers;
 	uint32 kmer_len;
 	uint32 minimizerVersion;
-	uint32 delta_val;
+	double delta_val;
 	uint32 window_len; 
 public:
 	CWSmallKSplitter(CKMCParams &Params, CKMCQueues &Queues);
